@@ -36,52 +45,67 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.6",
-    "@angular/common": "5.2.6",
-    "@angular/compiler": "5.2.6",
-    "@angular/core": "5.2.6",
-    "@angular/forms": "5.2.6",
-    "@angular/platform-browser": "5.2.6",
-    "@angular/platform-browser-dynamic": "5.2.6",
-    "@angular/router": "5.2.6",
-    "@ngrx/effects": "5.1.0",
-    "@ngrx/router-store": "5.0.1",
-    "@ngrx/store": "5.1.0",
-    "@ngrx/store-devtools": "5.1.0",
+    "@angular/animations": "13.0.0",
+    "@angular/common": "13.0.0",
+    "@angular/compiler": "13.0.0",
+    "@angular/core": "13.0.0",
+    "@angular/forms": "13.0.0",
+    "@angular/platform-browser": "13.0.0",
+    "@angular/platform-browser-dynamic": "13.0.0",
+    "@angular/router": "13.0.0",
+    "@ngrx/effects": "13.0.1",
+    "@ngrx/router-store": "13.0.1",
+    "@ngrx/store": "13.0.1",
+    "@ngrx/store-devtools": "13.0.1",
+    "@nrwl/angular": "13.2.2",
     "@nrwl/nx": "0.9.1",
     "angular-star-rating": "^4.0.0-beta.0",
     "bootstrap": "^4.0.0",
-    "core-js": "^2.4.1",
+    "core-js": "3.8.3",
     "css-star-rating": "^1.2.4",
     "fontawesome": "^4.7.2",
     "ngrx-store-freeze": "^0.2.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.6.7",
+    "rxjs-compat": "^6.0.0-rc.0",
+    "tslib": "^2.3.1",
+    "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "copyfiles":"2.0.0",
-    "@angular/cli": "1.7.1",
-    "@angular/compiler-cli": "5.2.6",
-    "@angular/language-service": "5.2.6",
-    "@nrwl/schematics": "0.9.1",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "@angular-devkit/build-angular": "13.0.3",
+    "@angular/cli": "13.0.3",
+    "@angular/compiler-cli": "13.0.0",
+    "@angular/language-service": "13.0.0",
+    "@nrwl/cli": "13.2.2",
+    "@nrwl/cypress": "13.2.2",
+    "@nrwl/eslint-plugin-nx": "13.2.2",
+    "@nrwl/jest": "13.2.2",
+    "@nrwl/linter": "13.2.2",
+    "@nrwl/schematics": "8.12.11",
+    "@nrwl/tao": "13.2.2",
+    "@nrwl/workspace": "13.2.2",
+    "@types/jasmine": "3.6.3",
+    "@types/jasminewd2": "2.0.6",
+    "@types/node": "14.14.33",
+    "codelyzer": "6.0.2",
     "compare-func": "^1.3.2",
+    "copyfiles": "2.4.1",
+    "cosmiconfig": "^7.0.1",
+    "cypress": "^8.3.1",
     "github-url-from-git": "^1.5.0",
+    "jasmine": "^3.8.0",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "~6.3.4",
+    "karma-chrome-launcher": "3.1.0",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "2.4.1",
-    "prettier": "1.10.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "2.6.2"
+    "karma-jasmine": "4.0.0",
+    "karma-jasmine-html-reporter": "1.7.0",
+    "ng-packagr": "13.0.7",
+    "postcss": "^8.3.6",
+    "prettier": "2.2.1",
+    "protractor": "~7.0.0",
+    "ts-node": "9.1.1",
+    "tslint": "~6.1.0",
+    "typescript": "4.4.4"
   }
 }
